debian: remove the git-arch package
https://www.gnu.org/software/gnu-arch/ states:

  GNU arch has been decommissioned. It has been superceded by
  GNU Bazaar.

The Debian package for git-arch may have been useful for people trying
to migrate from tla, but ten years is long enough for that.  If
git-archimport did not depend on tla then perhaps it could have lived
longer.  As it is, the maintenance burden of relying on such long
unmaintained software to support a very small potential set of users
is not worth it.

Thanks to Adrian Bunk for the suggestion.

Fixes: https://bugs.debian.org/866059
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>