debian: clarify choice between git-daemon-run, git-daemon-sysvinit
Standards-Version 4.0.1 eliminates the "extra" priority. That means
git-daemon-sysvinit get priority "optional", removing the illusion
that we've provided enough information for administrators to pick
which git-daemon service to install via the priority field.

Add some text to allow administrators to make an informed choice.
We should have done this long ago.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>