debian/gitweb: Recommends: lynx instead of lynx-cur
In 2015 Debian switched to packaging only one version of lynx.  The
lynx package is the package to depend on to get lynx and lynx-cur is a
transitional package that would eventually be removed.

Noticed by Łukasz 'sil2100' Zemczak in Ubuntu.  The report was not
forwarded to Debian but I stumbled upon it using the diff against
Ubuntu's package.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>