debian: add Breaks relation to help dgit upgrades
In support of a more predictable mapping from Git commits to Debian
source packages, dgit uses .git/info/attributes to disable handling of
attributes that affect the content in the worktree.  To support this,
dgit has a hard-coded list of Git attributes that can affect worktree
content.

In v2.18.0-rc0~95^2~3 (convert: add 'working-tree-encoding',
2018-04-15), Git gained a new attribute in this category.  Dgit
5099a427 (dgit: gitattributes: Defuse working-tree-encoding,
2018-06-20) adds this attribute to that hard-coded list to prevent
dgit from being affected by it.

Add a Breaks relation to ensure dgit is upgraded to a version that
guards against the working-tree-encoding attribute before Git is
upgraded to a version respecting it.

Fixes: https://bugs.debian.org/901897
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>