debian/rules: add support for DEB_BUILD_OPTIONS=terse
Debian policy 4.2.0.0 adds a new "terse" option for DEB_BUILD_OPTIONS
that can be used to request a quieter build (that still shows
progress).  More precisely, this allows

- showing the commands run by "make" in an abbreviated form
- suppressing irrelevant output from tests

This can be useful when building interactively since it makes it
easier to see at a glance how far the build has progressed.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>