debian/rules: use dpkg-architecture, not uname, to get arch
The "git version --build-options" command prints what arch Git was
built for.  Unfortunately, the upstream code to do this gets the arch
using the "uname -m" command instead of "cc -dumpmachine".
Fortunately, in the Debian environment we can get a nice host CPU
string using the "dpkg-architecture" command.

Noticed by the reproducible-builds continuous builder[1].  Thanks to
Ceridwen for adding this variation to reprotest in 0.1~15 ("Add
variations, notes theron, and tests", 2016-06-07).

[1] https://tests.reproducible-builds.org/debian/rb-pkg/unstable/i386/diffoscope-results/git.html

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>