debian/control: let git-all Suggest instead of Recommending git-daemon-run
Some users report that installing git-daemon-run causes systemd-sysv
to be deinstalled.  That is a problem in itself worth addressing, but
in the meantime it's also a chance to reevaluate why people are
installing git-daemon-run in the first place.  Most people installing
git-all are looking for gitk, git-svn, and similar tools and do not
need a server to be automatically launched as a side effect, so let's
downgrade its Recommends on git-daemon-run to a Suggests.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>