debian/git-el: do not exclude "emacs" from the flavors to install for
As Zack Weinberg noticed, the git-el package has not installed any
elisp for emacs for the last two years, producing a warning at
startup instead:

  git removed but not purged, skipping setup

That is because "emacs" switched from being a virtual package to
being the only flavor of standard emacs installed.  Update the
installation scripts to reflect this.

Even after this patch, the elisp installed does not do anything other
than point the user to magit and emacs's built in vc-git support.  A
followup change may remove the package.

Fixes: https://bugs.debian.org/972871
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>